Add script to write zeros to a block
If the power gets cut-off to a sata disk the checksum may not get
updated even though the block has been. This requires zeroing the block
to update the checksum so new data can be written to the block and
prevent read errors and failed smart tests on the disk.